requirements.txt
spacy
llama-index-llms-litellm
llama-index
python-dotenv